Python_lb/requirements.txt
python-dotenv
httpx
fastapi
uvicorn
docker
httpx
python-dotenv
